rename ScreenCurtain.RootViewController to ScreenCurtainViewController
