+++ wire-ios/Wire-iOS/Sources/AppDelegate.swift
@@ -281,10 +281,10 @@ private extension AppDelegate {
     }
 
     private func createAppRootRouter(launchOptions: LaunchOptions) {
+
         guard let viewController = window?.rootViewController as? RootViewController else {
             fatalError("rootViewController is not of type RootViewController")
         }
-
         guard let sessionManager = createSessionManager(launchOptions: launchOptions) else {
             fatalError("sessionManager is not created")
         }
